requirements.txt
langchain
streamlit
chromadb
pypdf
python-dotenv
openai
tiktoken
#pysqlite3-binary